refactor: merge `cleanHTML()` into `utils.ts`
