+++ lib/utils.ts
@@ -14,3 +14,11 @@ export function valueUpdater<T extends Updater<any>>(updaterOrValue: T, ref: Ref
       ? updaterOrValue(ref.value)
       : updaterOrValue
 }
+
+export function cleanHTML(content: string): string {
+  if (!content)
+    return ''
+
+  return content.replace(/<script[^>]*>([\s\S]*?)<\/script>/gmi, '')
+    .replace(/<style[^>]*>([\s\S]*?)<\/style>/gmi, '')
+}
